try to pull

+++ src/kudu/security/test/mini_kdc.cc
@@ -18,23 +18,21 @@
 #include "kudu/security/test/mini_kdc.h"
 
 #include <csignal>
-#include <stdlib.h>
+#include <cstdlib>
 
-#include <limits>
 #include <map>
 #include <memory>
 #include <string>
+#include <utility>
 
 #include <glog/logging.h>
 
-#include "kudu/gutil/gscoped_ptr.h"
-#include "kudu/gutil/strings/numbers.h"
-#include "kudu/gutil/strings/split.h"
 #include "kudu/gutil/strings/strip.h"
 #include "kudu/gutil/strings/substitute.h"
 #include "kudu/util/env.h"
 #include "kudu/util/monotime.h"
 #include "kudu/util/path_util.h"
+#include "kudu/util/slice.h"
 #include "kudu/util/stopwatch.h"
 #include "kudu/util/subprocess.h"
 #include "kudu/util/test_util.h"
@@ -42,31 +40,34 @@
 using std::map;
 using std::string;
 using std::unique_ptr;
+using std::vector;
 using strings::Substitute;
 
 namespace kudu {
 
 string MiniKdcOptions::ToString() const {
-  return strings::Substitute("{ realm: $0, port: $1, data_root: $2 }", realm, port, data_root);
+  return strings::Substitute("{ realm: $0, data_root: $1, port: $2, "
+      "ticket_lifetime: $3, renew_lifetime: $4 }",
+      realm, data_root, port, ticket_lifetime, renew_lifetime);
 }
 
 MiniKdc::MiniKdc()
     : MiniKdc(MiniKdcOptions()) {
 }
 
-MiniKdc::MiniKdc(const MiniKdcOptions& options)
-    : options_(options) {
+MiniKdc::MiniKdc(MiniKdcOptions options)
+    : options_(std::move(options)) {
   if (options_.realm.empty()) {
     options_.realm = "KRBTEST.COM";
   }
   if (options_.data_root.empty()) {
     options_.data_root = JoinPathSegments(GetTestDataDirectory(), "krb5kdc");
   }
-  if (options_.renew_lifetime.empty()) {
-    options_.renew_lifetime = "7d";
-  }
   if (options_.ticket_lifetime.empty()) {
     options_.ticket_lifetime = "24h";
+  }
+  if (options_.renew_lifetime.empty()) {
+    options_.renew_lifetime = "7d";
   }
 }
 
@@ -80,7 +81,9 @@ map<string, string> MiniKdc::GetEnvVars() const {
   return {
     {"KRB5_CONFIG", JoinPathSegments(options_.data_root, "krb5.conf")},
     {"KRB5_KDC_PROFILE", JoinPathSegments(options_.data_root, "kdc.conf")},
-    {"KRB5CCNAME", JoinPathSegments(options_.data_root, "krb5cc")}
+    {"KRB5CCNAME", JoinPathSegments(options_.data_root, "krb5cc")},
+    // Enable the workaround for MIT krb5 1.10 bugs from krb5_realm_override.cc.
+    {"KUDU_ENABLE_KRB5_REALM_FIX", "yes"}
   };
 }
 
@@ -97,34 +100,6 @@ vector<string> MiniKdc::MakeArgv(const vector<string>& in_argv) {
 
 namespace {
 // Attempts to find the path to the specified Kerberos binary, storing it in 'path'.
-Status GetBinaryPath(const string& binary,
-                     const vector<string>& search,
-                     string* path) {
-  string p;
-
-  // First, check specified locations which are sometimes not on the PATH.
-  // This is necessary to check first so that the system Heimdal kerberos
-  // binaries won't be found first on OS X.
-  for (const auto& location : search) {
-    p = JoinPathSegments(location, binary);
-    if (Env::Default()->FileExists(p)) {
-      *path = p;
-      return Status::OK();
-    }
-  }
-
-  // Next check if the binary is on the PATH.
-  Status s = Subprocess::Call({ "which", binary }, "", &p);
-  if (s.ok()) {
-    StripTrailingNewline(&p);
-    *path = p;
-    return Status::OK();
-  }
-
-  return Status::NotFound("Unable to find binary", binary);
-}
-
-// Attempts to find the path to the specified Kerberos binary, storing it in 'path'.
 Status GetBinaryPath(const string& binary, string* path) {
   static const vector<string> kCommonLocations = {
     "/usr/local/opt/krb5/sbin", // Homebrew
@@ -134,10 +109,9 @@ Status GetBinaryPath(const string& binary, string* path) {
     "/usr/lib/mit/sbin", // SLES
     "/usr/sbin", // Linux
   };
-  return GetBinaryPath(binary, kCommonLocations, path);
+  return FindExecutable(binary, kCommonLocations, path);
 }
 } // namespace
-
 
 Status MiniKdc::Start() {
   SCOPED_LOG_SLOW_EXECUTION(WARNING, 100, "starting KDC");
@@ -168,7 +142,7 @@ Status MiniKdc::Start() {
   RETURN_NOT_OK(GetBinaryPath("krb5kdc", &krb5kdc_bin));
 
   kdc_process_.reset(new Subprocess(
-      "env", MakeArgv({
+      MakeArgv({
       krb5kdc_bin,
       "-n", // Do not daemonize.
   })));
@@ -177,7 +151,7 @@ Status MiniKdc::Start() {
 
   const bool need_config_update = (options_.port == 0);
   // Wait for KDC to start listening on its ports and commencing operation.
-  RETURN_NOT_OK(WaitForKdcPorts());
+  RETURN_NOT_OK(WaitForUdpBind(kdc_process_->pid(), &options_.port, MonoDelta::FromSeconds(1)));
 
   if (need_config_update) {
     // If we asked for an ephemeral port, grab the actual ports and
@@ -190,7 +164,9 @@ Status MiniKdc::Start() {
 }
 
 Status MiniKdc::Stop() {
-  CHECK(kdc_process_);
+  if (!kdc_process_) {
+    return Status::OK();
+  }
   VLOG(1) << "Stopping KDC";
   unique_ptr<Subprocess> proc(kdc_process_.release());
   RETURN_NOT_OK(proc->Kill(SIGKILL));
@@ -204,6 +180,7 @@ Status MiniKdc::CreateKdcConf() const {
   static const string kFileTemplate = R"(
 [kdcdefaults]
 kdc_ports = $2
+kdc_tcp_ports = ""
 
 [realms]
 $1 = {
@@ -224,7 +201,7 @@ $1 = {
 Status MiniKdc::CreateKrb5Conf() const {
   static const string kFileTemplate = R"(
 [logging]
-    kdc = STDERR
+    kdc = FILE:/dev/stderr
 
 [libdefaults]
     default_realm = $1
@@ -233,6 +210,16 @@ Status MiniKdc::CreateKrb5Conf() const {
     forwardable = true
     renew_lifetime = $2
     ticket_lifetime = $3
+
+    # Disable aes256 since Java does not support it without JCE. Java is only
+    # one of several minicluster consumers, but disabling aes256 doesn't
+    # appreciably hurt Kudu code coverage, so we disable it universally.
+    #
+    # For more details, see:
+    # https://docs.oracle.com/javase/8/docs/technotes/guides/security/jgss/jgss-features.html
+    default_tkt_enctypes = aes128-cts des3-cbc-sha1
+    default_tgs_enctypes = aes128-cts des3-cbc-sha1
+    permitted_enctypes = aes128-cts des3-cbc-sha1
 
     # In miniclusters, we start daemons on local loopback IPs that
     # have no reverse DNS entries. So, disable reverse DNS.
@@ -257,67 +244,6 @@ Status MiniKdc::CreateKrb5Conf() const {
                                              options_.renew_lifetime, options_.ticket_lifetime);
   return WriteStringToFile(Env::Default(), file_contents,
                            JoinPathSegments(options_.data_root, "krb5.conf"));
-}
-
-Status MiniKdc::WaitForKdcPorts() {
-  SCOPED_LOG_SLOW_EXECUTION(WARNING, 100, Substitute("waiting for KDC ports"));
-  // We have to use 'lsof' to figure out which ports the KDC bound to if we
-  // requested ephemeral ones. The KDC doesn't log the bound port or expose it
-  // in any other fashion, and re-implementing lsof involves parsing a lot of
-  // files in /proc/. So, requiring lsof for tests and parsing its output seems
-  // more straight-forward. We call lsof in a loop in case the kdc is slow to
-  // bind to the ports.
-
-  string lsof;
-  RETURN_NOT_OK(GetBinaryPath("lsof", {"/sbin", "/usr/sbin"}, &lsof));
-
-  const vector<string> cmd = {
-    lsof, "-wbnP", "-Ffn",
-    "-p", std::to_string(kdc_process_->pid()),
-    "-a", "-i", "4UDP"};
-
-  string lsof_out;
-  for (int i = 1; ; i++) {
-    lsof_out.clear();
-    Status s = Subprocess::Call(cmd, "", &lsof_out);
-
-    if (s.ok()) {
-      StripTrailingNewline(&lsof_out);
-      break;
-    } else if (i > 10) {
-      return s;
-    }
-
-    SleepFor(MonoDelta::FromMilliseconds(i * i));
-  }
-
-  // The '-Ffn' flag gets lsof to output something like:
-  //   p19730
-  //   f123
-  //   n*:41254
-  // The first line is the pid. We ignore it.
-  // The second line is the file descriptor number. We ignore it.
-  // The third line has the bind address and port.
-  vector<string> lines = strings::Split(lsof_out, "\n");
-  int32_t port = -1;
-  if (lines.size() != 3 ||
-      lines[2].substr(0, 3) != "n*:" ||
-      !safe_strto32(lines[2].substr(3), &port) ||
-      port <= 0) {
-    return Status::RuntimeError("unexpected lsof output", lsof_out);
-  }
-  CHECK(port > 0 && port < std::numeric_limits<uint16_t>::max())
-      << "parsed invalid port: " << port;
-  VLOG(1) << "Determined bound KDC port: " << port;
-  if (options_.port == 0) {
-    options_.port = port;
-  } else {
-    // Sanity check: if KDC's port is already established, it's supposed to be
-    // written into the configuration files, so the process must bind to the
-    // already established port.
-    CHECK(options_.port == port);
-  }
-  return Status::OK();
 }
 
 Status MiniKdc::CreateUserPrincipal(const string& username) {
